feat: #94 종목 뉴스 조회 로직 구현

+++ back/news/src/main/java/com/ssafy/news/domain/repository/StockNewsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface StockNewsRepository extends JpaRepository<StockNews, Integer> {
+public interface StockNewsRepository extends JpaRepository<StockNews, Long> {
     @Query("select s from StockNews s order by s.uploadDatetime desc")
     Page<StockNews> findAllStockNews(Pageable pageable);
